more data source methods, insert and delete commands, zones and buttons

+++ app/src/main/java/com/sourceallies/android/zonebeacon/data/model/Gateway.java
@@ -41,7 +41,7 @@ public class Gateway implements DatabaseTable {
 
     @Setter
     @Getter
-    private int id;
+    private long id;
 
     @Setter
     @Getter
@@ -85,7 +85,7 @@ public class Gateway implements DatabaseTable {
             String column = cursor.getColumnName(i);
 
             if (column.equals(COLUMN_ID)) {
-                setId(cursor.getInt(i));
+                setId(cursor.getLong(i));
             } else if (column.equals(COLUMN_NAME)) {
                 setName(cursor.getString(i));
             } else if (column.equals(COLUMN_IP_ADDRESS)) {
