+++ app/src/main/java/com/sourceallies/android/zonebeacon/data/model/SystemType.java
@@ -34,7 +34,7 @@ public class SystemType implements DatabaseTable {
 
     @Setter
     @Getter
-    private int id;
+    private long id;
 
     @Setter
     @Getter
@@ -76,7 +76,7 @@ public class SystemType implements DatabaseTable {
             String column = cursor.getColumnName(i);
 
             if (column.equals(COLUMN_ID)) {
-                setId(cursor.getInt(i));
+                setId(cursor.getLong(i));
             } else if (column.equals(COLUMN_NAME)) {
                 setName(cursor.getString(i));
             } else if (column.equals(COLUMN_VERSION)) {
