+++ app/src/main/java/com/sourceallies/android/zonebeacon/data/model/ZoneButtonLink.java
@@ -37,7 +37,7 @@ public class ZoneButtonLink implements DatabaseTable {
 
     @Setter
     @Getter
-    private int id;
+    private long id;
 
     @Setter
     @Getter
@@ -73,7 +73,7 @@ public class ZoneButtonLink implements DatabaseTable {
             String column = cursor.getColumnName(i);
 
             if (column.equals(COLUMN_ID)) {
-                setId(cursor.getInt(i));
+                setId(cursor.getLong(i));
             } else if (column.equals(COLUMN_ZONE_ID)) {
                 setZoneId(cursor.getInt(i));
             } else if (column.equals(COLUMN_BUTTON_ID)) {
